Substantial refactoring of what's in ZIP-32 vs. the core spec
