+++ src/Nerdbank.Zcash/Zip32HDWallet.Sapling.ExtendedSpendingKey.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Andrew Arnott. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Numerics;
+using Nerdbank.Zcash.Sapling;
 
 namespace Nerdbank.Zcash;
 
@@ -41,6 +41,9 @@ public partial class Zip32HDWallet
 			/// Gets the extended full viewing key.
 			/// </summary>
 			public ExtendedFullViewingKey FullViewingKey => this.fullViewingKey ??= new(this);
+
+			/// <inheritdoc/>
+			public FullViewingKeyFingerprint Fingerprint => GetFingerprint(this.FullViewingKey.Key);
 
 			/// <inheritdoc/>
 			public FullViewingKeyTag ParentFullViewingKeyTag { get; }
